Refactor: Remplacer le composant ChannelList par une récupération des canaux dans le composant Drawer et mettre à jour les icônes de déconnexion
