+++ client/src/components/navigation/app/Drawer.tsx
@@ -1,19 +1,56 @@
-import { JSX } from "react";
+import { JSX, useEffect, useState } from "react";
+import { useChatStore } from "../../../store/chatStore";
+import axios from "axios";
 
 interface DrawerProps {
     isOpen: boolean;
 }
 
 export default function Drawer({ isOpen }: DrawerProps): JSX.Element {
+    const { channels, setChannels, joinChannel, currentChannel } = useChatStore();
+    const [loading, setLoading] = useState(false);
+    const API_URL = import.meta.env.VITE_API_URL;
+
+    useEffect(() => {
+        const fetchChannels = async () => {
+            try {
+                setLoading(true);
+                const response = await axios.get(`${API_URL}/api/channels`, {
+                    withCredentials: true,
+                });
+                setChannels(response.data);
+            } catch (error) {
+                console.error("Failed to fetch channels:", error);
+            } finally {
+                setLoading(false);
+            }
+        };
+
+        fetchChannels();
+    }, [setChannels]);
+
+    if (loading) return <div>Loading channels...</div>;
+
     return (
-        <div className={`drawer ${isOpen ? "drawer-open" : ""} w-0`}>
+        <div className={`drawer ${isOpen ? "drawer-open w-full max-w-xs" : "w-0"} `}>
             <input id="my-drawer" type="checkbox" className="drawer-toggle" checked={isOpen} readOnly />
-            <div className="drawer-side transition-transform duration-300 ease-out transform translate-x-0">
+            <div className="drawer-side">
                 <label htmlFor="my-drawer" aria-label="close sidebar" className="drawer-overlay"></label>
                 <ul className="menu bg-base-200 text-base-content min-h-full w-80 p-4">
-                    {/* Sidebar content here */}
-                    <li><a>Sidebar Item 1</a></li>
-                    <li><a>Sidebar Item 2</a></li>
+                    <li>
+                        <h2 className="menu-title">Channels</h2>
+                        <ul>
+                            {channels.map((channel) => (
+                                <li
+                                    key={channel._id}
+                                    className={channel._id === currentChannel ? "active" : ""}
+                                    onClick={() => channel._id && joinChannel(channel._id)}
+                                >
+                                    {channel.name}
+                                </li>
+                            ))}
+                        </ul>
+                    </li>
                 </ul>
             </div>
         </div>
