+++ client/src/pages/app/ChatPage.tsx
@@ -1,10 +1,9 @@
 // src/pages/ChatPage.tsx
 import { useEffect } from "react";
-import { ChannelList } from "../../components/channel/ChannelList";
 import { Chat } from "../../components/chat/Chat";
 import { SocketEventHandler } from "../../components/socket/SocketEventHandler";
 import { useSocketStore } from "../../socket/socket";
-import {useAuth} from "../../context/AuthContext.tsx";
+import { useAuth } from "../../context/AuthContext.tsx";
 
 export const ChatPage = () => {
     const { connect, disconnect } = useSocketStore();
@@ -26,9 +25,6 @@ export const ChatPage = () => {
         <div className="chat-page">
             <SocketEventHandler />
             <div className="chat-layout">
-                <aside className="sidebar">
-                    <ChannelList />
-                </aside>
                 <main className="main-content">
                     <Chat userId={user.userId} username={user.username} />
                 </main>
